dun_render: Avoid `Invalid` tile type
1. Do not call `RenderTile` with invalid values to begin with.
2. Move mask selection out of `RenderTile`.

+++ Source/levels/gendung.cpp
@@ -34,7 +34,7 @@ bool setlevel;
 _setlevels setlvlnum;
 dungeon_type setlvltype;
 Point ViewPosition;
-int MicroTileLen;
+uint_fast8_t MicroTileLen;
 char TransVal;
 bool TransList[256];
 uint16_t dPiece[MAXDUNX][MAXDUNY];
